docs(repository): Configure repository labels in code (#10)
Configure the repository labels in code, for usage in a github context.

This allows the labels of a repository to be controlled within the repository. In combination with diff analysis, changes to the codebase as a whole can be automatically labelled in pull requests and issues. For now this just aims to set a standard in all of the repositories, so the labels are somewhat generic.